Move Resolver to a legacy_resolve module
